Refactor biodynamo cli to be more modular.
